Removed NarrationTextItemData and replaced it with NarratableItemData
